Fixed issue in Prins encoding (#1177)
git-svn-id: https://src.heuristiclab.com/svn/core/trunk@6758 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ sources/HeuristicLab.Problems.VehicleRouting/3.3/Encodings/Prins/PrinsEncoding.cs
@@ -85,7 +85,7 @@ namespace HeuristicLab.Problems.VehicleRouting.Encodings.Prins {
       double[] V = new double[cities + 1];
       V[0] = 0;
       for (int i = 1; i <= cities; i++) {
-        V[i] = int.MaxValue;
+        V[i] = double.MaxValue;
       }
 
       for (int i = 1; i <= cities; i++) {
